auto aim using turret
